1129commit

+++ finalProject/src/main/java/com/example/finalProject/controller/HomeController.java
@@ -88,7 +88,7 @@ public class HomeController {
 
 		try {
 			// 사용자 정보 가져오기
-			Long userId = memberService.getAutoIncrementIdByEmail(user);
+			int userId = memberService.getAutoIncrementIdByEmail(user);
 			String userName = memberService.getUsernameByEmail(user);
 			String prompt = memberService.getPromptByEmail(user);
